fix(input): 修复依赖路径问题
