+++ demo/components.d.ts
@@ -19,21 +19,17 @@ declare module 'vue' {
     TopBar: typeof import('./src/components/TopBar.vue')['default']
     Yikedesign: typeof import('./src/components/design/yikedesign.vue')['default']
     Yikedev: typeof import('./src/components/develops/yikedev.vue')['default']
-    YkAlert: typeof import('./../packages/yike-design-ui/src/index.ts')['YkAlert']
-    YkAnchor: typeof import('./../packages/yike-design-ui/src/index.ts')['YkAnchor']
-    YkAvatar: typeof import('./../packages/yike-design-ui/src/index.ts')['YkAvatar']
-    YkBackTop: typeof import('./../packages/yike-design-ui/src/index.ts')['YkBackTop']
-    YkBadge: typeof import('./../packages/yike-design-ui/src/index.ts')['YkBadge']
-    YkButton: typeof import('./../packages/yike-design-ui/src/index.ts')['YkButton']
-    YkContainer: typeof import('./../packages/yike-design-ui/src/index.ts')['YkContainer']
-    YkDrawer: typeof import('./../packages/yike-design-ui/src/index.ts')['YkDrawer']
-    YkIcon: typeof import('./../packages/yike-design-ui/src/index.ts')['YkIcon']
-    YkInput: typeof import('./../packages/yike-design-ui/src/index.ts')['YkInput']
-    YkInputSearch: typeof import('./../packages/yike-design-ui/src/index.ts')['YkInputSearch']
-    YkParagraph: typeof import('./../packages/yike-design-ui/src/index.ts')['YkParagraph']
-    YkSpace: typeof import('./../packages/yike-design-ui/src/index.ts')['YkSpace']
-    YkText: typeof import('./../packages/yike-design-ui/src/index.ts')['YkText']
-    YkTheme: typeof import('./../packages/yike-design-ui/src/index.ts')['YkTheme']
-    YkTitle: typeof import('./../packages/yike-design-ui/src/index.ts')['YkTitle']
+    YkAvatar: typeof import('../packages/yike-design-ui/src/index.ts')['YkAvatar']
+    YkBackTop: typeof import('../packages/yike-design-ui/src/index.ts')['YkBackTop']
+    YkButton: typeof import('../packages/yike-design-ui/src/index.ts')['YkButton']
+    YkDrawer: typeof import('../packages/yike-design-ui/src/index.ts')['YkDrawer']
+    YkIcon: typeof import('../packages/yike-design-ui/src/index.ts')['YkIcon']
+    YkInput: typeof import('../packages/yike-design-ui/src/index.ts')['YkInput']
+    YkInputSearch: typeof import('../packages/yike-design-ui/src/index.ts')['YkInputSearch']
+    YkScrollbar: typeof import('../packages/yike-design-ui/src/index.ts')['YkScrollbar']
+    YkSpace: typeof import('../packages/yike-design-ui/src/index.ts')['YkSpace']
+    YkText: typeof import('../packages/yike-design-ui/src/index.ts')['YkText']
+    YkTheme: typeof import('../packages/yike-design-ui/src/index.ts')['YkTheme']
+    YkTitle: typeof import('../packages/yike-design-ui/src/index.ts')['YkTitle']
   }
 }
