+++ packages/yike-design-ui/src/components/input-search/index.ts
@@ -1,0 +1,4 @@
+import InputSearch from './src/input-search.vue';
+import { withInstall } from '../utils';
+export const YkInputSearch = withInstall(InputSearch);
+export * from './src/input-search';
