+++ packages/yike-design-ui/src/components/input-search/src/input-search.ts
@@ -1,0 +1,10 @@
+import { Size, Status } from '../../utils/constant';
+
+export type InputSearchProps = {
+  name?: string;
+  id?: string;
+  size?: Size;
+  placeholder?: string;
+  clearable?: boolean;
+  loading?: boolean;
+};
