+++ packages/yike-design-ui/src/components/input-search/style/index.less
@@ -1,0 +1,1 @@
+@import '../../styles/color/colors.less';
