+++ packages/yike-design-ui/src/components/input-search/style/index.ts
@@ -1,0 +1,1 @@
+import './index.less';
