+++ packages/yike-design-ui/src/components/input/index.ts
@@ -1,6 +1,4 @@
 import Input from './src/input.vue';
-import InputSearch from './src/input-search.vue';
-import { withInstall } from '../../utils/index';
+import { withInstall } from '../utils/index';
 export const YkInput = withInstall(Input);
-export const YkInputSearch = withInstall(InputSearch);
 export * from './src/input';
