+++ packages/yike-design-ui/src/components/input/style/index.less
@@ -1,3 +1,5 @@
+@import '../../styles/color/colors.less';
+
 .yk-input {
   display: inline-flex;
   flex-basis: content;
