+++ packages/yike-design-ui/src/index.ts
@@ -10,7 +10,8 @@ import { YkDrawer } from './components/drawer';
 import YkBadge from './components/badge';
 import YkScrollbar from './components/scrollbar';
 import YkRate from './components/rate';
-import { YkInput, YkInputSearch } from './components/input';
+import { YkInput } from './components/input';
+import { YkInputSearch } from './components/input-search';
 import YkAlert from './components/alert';
 import YkButton from './components/button';
 import YkTheme from './components/theme';
